% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{compileAndSaveAllResults}
\alias{compileAndSaveAllResults}
\title{Compile the results of a simulation + classifier methods run}
\usage{
compileAndSaveAllResults(run.results = NULL, save.file = NULL,
  verbose = FALSE)
}
\arguments{
\item{run.results}{A list of run results}

\item{save.file}{A character vector for results filename or NULL to skip}

\item{verbose}{A flag indicating whether verbose output be sent to stdout}
}
\value{
A list with:
\describe{
  \item{algo.acc}{data frame of results, a row for each update}
  \item{ggplot.data}{melted results data frame for plotting}
  \item{correct}{number of variables detected correctly in each data set}
}
}
\description{
Compile the results of a simulation + classifier methods run
}
\seealso{
Other workflows: \code{\link{paperRealWorkflow}},
  \code{\link{paperSimWorkflow}},
  \code{\link{plotRunResults}}
}
