% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{createInteractions}
\alias{createInteractions}
\title{Create a differentially coexpressed data set with interactions}
\usage{
createInteractions(M = 100, N = 100, class.label = "class",
  meanExpression = 7, A = NULL, randSdNoise = 1, sdNoise = 0.4,
  sampleIndicesInteraction = NULL, verbose = FALSE)
}
\arguments{
\item{M}{An integer for the number of samples (rows)}

\item{N}{An integer for the number of variables (columns)}

\item{class.label}{A character vector for the name of the class column}

\item{meanExpression}{A numeric for the mean expression levels}

\item{A}{A matrix representing a weighted, undirected network (adjacency)}

\item{randSdNoise}{Random noise for the background expression levels}

\item{sdNoise}{A numeric for the noise in the differential expression}

\item{sampleIndicesInteraction}{A vector of integers of significant variables}

\item{verbose}{A flag for sending verbose output to stdout}
}
\value{
A matrix representing the new new data set.
}
\description{
Create a differentially coexpressed data set with interactions
}
\seealso{
Other simulation: \code{\link{createMainEffects}},
  \code{\link{createSimulation}},
  \code{\link{splitDataset}}
}
