% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{createMixedSimulation}
\alias{createMixedSimulation}
\title{Create a data simulation with a mix of main and interaction effects.}
\usage{
createMixedSimulation(num.samples = 100, num.variables = 100,
  pct.imbalance = 0.5, pct.signals = 0.1, bias = 0.4, label = "class",
  pct.mixed = 0.5, mixed.type = c("mainEffect", "interactionScalefree"),
  pct.train = 0.5, pct.holdout = 0.5, pct.validation = 0,
  save.file = NULL, verbose = FALSE)
}
\arguments{
\item{num.samples}{An integer for the number of samples}

\item{num.variables}{An integer for the number of variables}

\item{pct.imbalance}{A numeric percentage to indicate proportion of the imbalaced samples. 
0 means all controls and 1 mean all cases.}

\item{pct.signals}{A numeric for proportion of simulated signal variables}

\item{bias}{A numeric for effect size in simulated signal variables}

\item{label}{A character vector for the name of the outcome column. class for classification
and phenos for regression}

\item{pct.mixed}{A numeric percentage to indicate the proportion of each simulation type}

\item{mixed.type}{A character vector of the mix type of simulations:
mainEffect, interactionErdos/interactionScalefree}

\item{pct.train}{A numeric percentage of samples to use for traning}

\item{pct.holdout}{A numeric percentage of samples to use for holdout}

\item{pct.validation}{A numeric percentage of samples to use for testing}

\item{save.file}{A filename or NULL indicating whether to save the simulations to file}

\item{verbose}{A flag indicating whether verbose output be sent to stdout}
}
\value{
A list with:
\describe{
  \item{train}{traing data set}
  \item{holdout}{holdout data set}
  \item{validation}{validation data set}
  \item{label}{the class label/column name}
  \item{signal.names}{the variable names with simulated signals}
  \item{elapsed}{total elapsed time}
}
}
\description{
Create a data simulation with a mix of main and interaction effects.
}
\examples{
num.variables <- 100
num.samples <- 100
pct.imbalance <- 0.5
pct.signals <- 0.1
bias <- 0.4
label <- "class"
pct.mixed <- 0.5
mixed.type <- c("mainEffect", "interactionScalefree")
sim.data <- createMixedSimulation(num.samples=num.samples,
                                  num.variables=num.variables,
                                  pct.signals=pct.signals,
                                  pct.imbalance=pct.imbalance,
                                  label = label,
                                  bias=bias,
                                  pct.mixed=pct.mixed,
                                  mixed.type=mixed.type,
                                  verbose=FALSE)
}
\seealso{
Other simulation: \code{\link{createInteractions}},
  \code{\link{createMainEffects}},
  \code{\link{createSimulation}},
  \code{\link{splitDataset}}
}
