% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{createSimulation}
\alias{createSimulation}
\title{Create a data simulation and return train/holdout/validation data sets.}
\usage{
createSimulation(n = 100, num.vars = 100, pct.signals = 0.1, bias = 0.4,
  class.label = "class", shortname = "paramstring",
  sim.type = "mainEffect", myrun = "001", verbose = FALSE,
  save.file = FALSE)
}
\arguments{
\item{n}{An integer for the number of samples}

\item{num.vars}{An integer for the number of variables}

\item{pct.signals}{A numeric for proportion of simulated signal variables}

\item{bias}{A numeric for effect size in simulated signal variables}

\item{class.label}{A character vector for the name of the class column}

\item{shortname}{A character vector of a parameters separated by '_'}

\item{sim.type}{A character vector of the type of simulation:
mainEffect/interactionErdos/interactionScalefree}

\item{myrun}{A character vector of a unique run identifier}

\item{verbose}{A flag indicating whether verbose output be sent to stdout}

\item{save.file}{A flag indicating whther to save the results to file}
}
\value{
A list with:
\describe{
  \item{train}{traing data set}
  \item{holdout}{holdout data set}
  \item{validation}{validation data set}
  \item{class.label}{the class label/column name}
  \item{signal.names}{the variable names with simulated signals}
  \item{elapsed}{total elapsed time}
}
}
\description{
Create a data simulation and return train/holdout/validation data sets.
}
\examples{
sim.type <- "mainEffect"
num.samples <- 100
num.variables <- 100
pct.signals <- 0.1
bias <- 0.4
sim.data <- createSimulation(num.vars=num.variables,
                             n=num.samples,
                             pct.signals=pct.signals,
                             bias=bias,
                             sim.type=sim.type,
                             verbose=FALSE)
}
\seealso{
Other simulation: \code{\link{createDiffCoexpMatrixNoME}},
  \code{\link{simulateData}}, \code{\link{splitDataset}}
}
