% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{epistasisRank}
\alias{epistasisRank}
\title{Compute and return epistasis rank scores}
\usage{
epistasisRank(G, Gamma_vec)
}
\arguments{
\item{G}{Am adjacency matrix}

\item{Gamma_vec}{A vector of prior knowledge}
}
\description{
Modified SNPrank function
}
\seealso{
Other classification: \code{\link{getImportanceScores}},
  \code{\link{originalThresholdout}},
  \code{\link{privateEC}}, \code{\link{privateRF}},
  \code{\link{standardRF}}, \code{\link{xgboostRF}}
}
