% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pecdata.R
\docType{data}
\name{fullfMRI2}
\alias{fullfMRI2}
\title{An MRI data set used used in the paper XXX.}
\format{Data frame with 80 rows and 3003 all-pairs variables.}
\usage{
data(fullfMRI2)
}
\description{
The fMRI data includes 80 unmedicated MDD (52 females, age±sd. 33±11) and 80
healthy controls (HCs) (41 females, age±sd. 31±10). We used AFNI(Cox, 1996)
to process the rs-fMRI data and extract 3003 z-transformed correlation
coefficients between 78 brain regions identified by a functional region of
interest atlas (Shirer, et al., 2012).
}
\keyword{datasets}
