% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{getImportanceScores}
\alias{getImportanceScores}
\title{Compute and return importance scores (Relief-F scores)}
\usage{
getImportanceScores(train.set = NULL, holdout.set = NULL, label = "class",
  importance.name = "relieff", importance.algorithm = "ReliefFbestK",
  verbose = FALSE)
}
\arguments{
\item{train.set}{A training data frame with last column as outcome}

\item{holdout.set}{A holdout data frame with last column as outcome}

\item{label}{A character vector of the outcome variable column name. class/phenos for classification/regression}

\item{importance.name}{A list importance operation parameters}

\item{importance.algorithm}{A character vector of the ReliefF estimator}

\item{verbose}{A flag indicating whether verbose output be sent to stdout}
}
\value{
A list with two data frames representing the importance scores
(Relief-F scores) for the train and holdout data sets.
}
\description{
Compute and return importance scores (Relief-F scores)
}
\seealso{
Other classification: \code{\link{epistasisRank}},
  \code{\link{originalThresholdout}},
  \code{\link{privateEC}}, \code{\link{privateRF}},
  \code{\link{standardRF}}, \code{\link{xgboostRF}}
}
