% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{originalPrivacy}
\alias{originalPrivacy}
\title{Original privacy algorithm}
\usage{
originalPrivacy(data.sets = NULL, is.simulated = TRUE, n = 100, d = 100,
  shortname = "paramstring", type = "sva", myrun = "001",
  update.freq = 50, pec.file = NULL, threshold = 4/sqrt(n),
  tolerance = 1/sqrt(n), signal.names = NULL, save.file = NULL,
  verbose = FALSE)
}
\arguments{
\item{data.sets}{A list of train, holdout and test data frames}

\item{is.simulated}{Is the data simulated (or real?)}

\item{n}{An integer for the number of samples}

\item{d}{An integer for the number of variables}

\item{shortname}{A character vector of a parameters separated by '_'}

\item{type}{A character vector of the type of simulation: sva|er|inte}

\item{myrun}{A character vector of a unique run identifier}

\item{update.freq}{A integer for the number of steps before update}

\item{pec.file}{A character vector filename of privateEC results}

\item{threshold}{A numeric, default 4 / sqrt(n) suggested in the
thresholdout’s supplementary material (Dwork, et al.,2015)}

\item{tolerance}{A numeric, default 1 / sqrt(n) suggested in the
thresholdout’s supplementary material (Dwork, et al.,2015)}

\item{signal.names}{A character vector of signal names in simulated data}

\item{save.file}{A character vector for results filename or NULL to skip}

\item{verbose}{A flag indicating whether verbose output be sent to stdout}
}
\value{
A list containing:
\describe{
  \item{plots.data}{data frame of results, a row for each update}
  \item{melted.data}{melted results data frame for plotting}
  \item{correct}{number of variables detected correctly in each data set}
  \item{elapsed}{total elapsed time}
}
}
\description{
Original Thresholdout with Dwork’s linear classifier
(Dwork, et al., 2015)
}
\examples{
num.samples <- 100
num.variables <- 100
pb <- 0.1
nbias <- pb * num.variables
signals <- sprintf("gene\%04d", 1:nbias)
temp.pec.file <- tempfile(pattern="pEc_temp", tmpdir=tempdir())

data.sets <- createSimulation(d=num.variables,
                              n=num.samples,
                              type="sva",
                              verbose=FALSE)
pec.results <- privateEC(data.sets=data.sets,
                         is.simulated=TRUE,
                         n=num.samples,
                         d=num.variables,
                         signal.names=signals,
                         save.file=temp.pec.file,
                         verbose=FALSE)
por.results <- originalPrivacy(data.sets=data.sets,
                               is.simulated=TRUE,
                               n=num.samples,
                               d=num.variables,
                               signal.names=signals,
                               pec.file=temp.pec.file,
                               verbose=FALSE)
file.remove(temp.pec.file)
}
\seealso{
Other classification: \code{\link{privateECinbix}},
  \code{\link{privateEC}}, \code{\link{standardRF}}
}
