% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{paperSimWorkflow}
\alias{paperSimWorkflow}
\title{Workflow for running one simulation of the Bioinformatics paper workflow}
\usage{
paperSimWorkflow(myrun = "001", n = 100, num.vars = 100,
  pct.signals = 0.1, update.freq = 50, verbose = FALSE)
}
\arguments{
\item{myrun}{A character vector identifying the run}

\item{n}{An integer for the number of samples}

\item{num.vars}{An integer for the number of variables}

\item{pct.signals}{A numeric for the significant variable bias}

\item{update.freq}{A integer for the number of steps before update}

\item{verbose}{A flag indicating whether verbose output be sent to stdout}
}
\value{
A list containing:
\describe{
  \item{run.results}{data frame of run results of each sim type}
  \item{elapsed}{total elapsed time}
}
}
\description{
Creates one simulation of train/holdout/validation data sets, then runs the
four algorithms on that data. Returns a data frame of run results for each.
}
\note{
Default parameter values match those from the Bioinformatics paper.
}
\examples{
  num.samples <- 100
  num.variables <- 100
  pct.signals <- 0.1
  bias <- 0.4
  one.step.result <- paperSimWorkflow(myrun="001", n=100, num.vars=100, pct.signals=0.1,
  update.freq=50, verbose=FALSE)
}
\seealso{
The workflow consists of the sequence:
\code{\link{createSimulation}}
\code{\link{privateEC}} (optionally \code{\link{privateECinbix}})
\code{\link{originalThresholdout}}
\code{\link{privateRF}}
\code{\link{standardRF}} and
\code{\link{compileResults}}. A comparison analysis with real data (fMRI)
is in \code{\link{paperRealWorkflow}}.

Other workflows: \code{\link{compileResults}},
  \code{\link{paperRealWorkflow}}
}
