% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pecdata.R
\docType{data}
\name{phenos}
\alias{phenos}
\title{An MRI data set phenotypes..}
\format{Vector of phenotype values as a factor.}
\usage{
data(phenos)
}
\description{
The fMRI data includes 80 unmedicated MDD (52 females, age±sd. 33±11) and 80
healthy controls (HCs) (41 females, age±sd. 31±10).
}
\keyword{datasets}
