% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{privateECinbix}
\alias{privateECinbix}
\title{Call C++ inbix Evaporative Cooling Privacy}
\usage{
privateECinbix(data.sets = NULL, is.simulated = TRUE, n = 100,
  shortname = "paramstring", bias = 0.4, type = "sva", myrun = "001",
  update.freq = 50, start.temp = 0.1, final.temp = 10^(-5),
  tau.param = 100, save.file = NULL, verbose = FALSE)
}
\arguments{
\item{data.sets}{A list of train, holdout and test data frames}

\item{is.simulated}{Is the data simulated (or real?)}

\item{n}{An integer for the number of samples}

\item{shortname}{A character vector of a parameters separated by '_'}

\item{bias}{A numeric for ?}

\item{type}{A character vector of the type of simulation: sva|er|inte}

\item{myrun}{A character vector of a unique run identifier}

\item{update.freq}{A integer for the number of steps before update}

\item{start.temp}{A numeric for EC starting temperature}

\item{final.temp}{A numeric for EC final temperature}

\item{tau.param}{A numeric for tau to control reduction schedule}

\item{save.file}{A character vector for results filename or NULL to skip}

\item{verbose}{A flag indicating whether verbose output be sent to stdout}
}
\value{
A list containing:
\describe{
  \item{plots.data}{data frame of results, a row for each update}
  \item{melted.data}{melted results data frame for plotting}
  \item{correct}{number of variables detected correctly in each data set}
  \item{elapsed}{total elapsed time}
}
}
\description{
Assumes the inbix executable is in the PATH.
}
\note{
inbix must be instaled in the path
}
\examples{
num.samples <- 100
num.variables <- 100
pb <- 0.1
nbias <- pb * num.variables
signals <- sprintf("gene\%04d", 1:nbias)
sim.data <- createSimulation(d=num.variables, n=num.samples,
                             type="sva", verbose=FALSE)
pec.results <- privateECinbix(data.sets=sim.data,
                              n=num.samples,
                              verbose=FALSE)
}
\seealso{
Other classification: \code{\link{originalPrivacy}},
  \code{\link{privateEC}}, \code{\link{standardRF}}
}
