% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{privateRF}
\alias{privateRF}
\title{Private random forests algorithm}
\usage{
privateRF(data.sets = NULL, is.simulated = TRUE, n = 100, d = 100,
  shortname = "paramstring", type = "sva",
  rf.importance.measure = "MeanDecreaseGini", pec.file = NULL,
  update.freq = 50, threshold = 4/sqrt(n), tolerance = 1/sqrt(n),
  signal.names = NULL, save.file = NULL, verbose = FALSE)
}
\arguments{
\item{data.sets}{A list of train, holdout and test data frames}

\item{is.simulated}{Is the data simulated (or real?)}

\item{n}{An integer for the number of samples}

\item{d}{An integer for the number of variables}

\item{shortname}{A character vector of a parameters separated by '_'}

\item{type}{A character vector of the type of simulation: sva|er|inte}

\item{rf.importance.measure}{A character vector for the random forest importance measure}

\item{pec.file}{A character vector filename of privateEC results}

\item{update.freq}{A integer for the number of steps before update}

\item{threshold}{A numeric, default 4 / sqrt(n) suggested in the
thresholdout’s supplementary material (Dwork, et al.,2015)}

\item{tolerance}{A numeric, default 1 / sqrt(n) suggested in the
thresholdout’s supplementary material (Dwork, et al.,2015)}

\item{signal.names}{A character vector of signal names in simulated data}

\item{save.file}{A character vector for results filename or NULL to skip}

\item{verbose}{A flag indicating whether verbose output be sent to stdout}
}
\value{
A list containing:
\describe{
  \item{plots.data}{data frame of results, a row for each update}
  \item{melted.data}{melted results data frame for plotting}
  \item{correct}{number of variables detected correctly in each data set}
  \item{elapsed}{total elapsed time}
}
}
\description{
Random Forest Thresholdout, which is TO with the feature selection
and classifier replaced with Random Forest.
}
\examples{
num.samples <- 100
num.variables <- 100
pb <- 0.1
nbias <- pb * num.variables
signals <- sprintf("gene\%04d", 1:nbias)
temp.pec.file <- tempfile(pattern="pEc_temp", tmpdir=tempdir())

data.sets <- createSimulation(d=num.variables,
                              n=num.samples,
                              type="sva",
                              verbose=FALSE)
pec.results <- privateEC(data.sets=data.sets,
                         is.simulated=TRUE,
                         n=num.samples,
                         d=num.variables,
                         signal.names=signals,
                         save.file=temp.pec.file,
                         verbose=FALSE)
prf.results <- privateRF(data.sets=data.sets,
                         is.simulated=TRUE,
                         n=num.samples,
                         d=num.variables,
                         signal.names=signals,
                         pec.file=temp.pec.file,
                         verbose=FALSE)
file.remove(temp.pec.file)
}
