% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{saveSimAsInbixNative}
\alias{saveSimAsInbixNative}
\title{Write inbix numeric and phenotype files (PLINK format)}
\usage{
saveSimAsInbixNative(data.sets = NULL, base.sim.prefix, verbose = FALSE)
}
\arguments{
\item{data.sets}{A list of train, holdout and test data frames}

\item{base.sim.prefix}{A character vector for the input and saved file prefixes}

\item{verbose}{A flag for sending berbose output to stdout}
}
\description{
Write inbix numeric and phenotype files (PLINK format)
}
