% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulateData}
\alias{simulateData}
\title{Create a simulated data set}
\usage{
simulateData(n.e = 1000, n.db = 70, n.ns = 30, sv.db = c("A", "B"),
  sv.ns = c("A", "B"), sd.b = 1, sd.gam = 1, sd.u = 1, conf = FALSE,
  distr.db = NA, p.b = 0.3, p.gam = 0.3, p.ov = p.b/2)
}
\arguments{
\item{n.e}{number of variables}

\item{n.db}{sample size in database}

\item{n.ns}{sample size in newsample}

\item{sv.db}{batches}

\item{sv.ns}{batches}

\item{sd.b}{a numeric for sd.b?}

\item{sd.gam}{a numeric for sd.gam?}

\item{sd.u}{a numeric for sd.u?}

\item{conf}{a flag for conf?}

\item{distr.db}{a numeric for distr.db?}

\item{p.b}{a numeric for p.b?}

\item{p.gam}{a numeric for p.gam?}

\item{p.ov}{a numeric for p.ov?}
}
\value{
A list with:
\describe{
  \item{db}{database creation variables}
  \item{vars}{variables used in simulation}
}
}
\description{
simulating \eqn{X = BS + \Gamma G + U}
}
\details{
S = Biological group                                                                                                                   m
G = Batch
U = random error

NOTE:  If you use conf=TRUE, then you must have exactly two surrogate
variables in the database this function only allows for confounding in
the database, not confounding in the new samples
}
\seealso{
Other simulation: \code{\link{createDiffCoexpMatrixNoME}},
  \code{\link{createSimulation}},
  \code{\link{splitDataset}}
}
