% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{splitDataset}
\alias{splitDataset}
\title{Split a data set for machine learning classification}
\usage{
splitDataset(all.data = NULL, pct.train = 0.5, pct.holdo = 0.5,
  pct.validation = 0, class.label = "phenos")
}
\arguments{
\item{all.data}{A data frame of n rows by d colums of data plus a label column}

\item{pct.train}{A numeric percentage of samples to use for traning}

\item{pct.validation}{A numeric percentage of samples to use for testing}

\item{class.label}{A character vector of the data column name for the class label}

\item{pct.holdout}{A numeric percentage of samples to use for holdout}
}
\value{
A list containing:
\describe{
  \item{train}{traing data set}
  \item{holdout}{holdout data set}
  \item{validation}{validation data set}
}
}
\description{
Return data.sets as a list of training set, holdout set and validation set
according to the predefined percentage of each partition
default is a 50-50 split into training and holdout, no testing set
code class/label/phenotypes as 1 and -1.
}
\examples{
data("fullfMRI2")
data.sets <- splitDataset(fullfMRI2)
}
\seealso{
Other simulation: \code{\link{createDiffCoexpMatrixNoME}},
  \code{\link{createSimulation}},
  \code{\link{simulateData}}
}
