% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{createSimulation}
\alias{createSimulation}
\title{Create a data simulation and return train/holdout/test data sets.}
\usage{
createSimulation(n = 100, d = 100, pb = 0.1, bias = 0.4,
  shortname = "paramstring", type = "sva", myrun = "001",
  verbose = FALSE, save.file = FALSE)
}
\arguments{
\item{n}{An integer for the number of samples}

\item{d}{An integer for the number of variables}

\item{pb}{A numeric for proportion of functional variables}

\item{bias}{A numeric for bias in data simulation}

\item{shortname}{A character vector of a parameters separated by '_'}

\item{type}{A character vector of the type of simulation: sva|er|inte}

\item{myrun}{A character vector of a unique run identifier}

\item{verbose}{A flag indicating whether verbose output be sent to stdout}

\item{save.file}{A flag indicating whther to save the results to file}
}
\value{
A list with:
\describe{
  \item{plots.data}{data frame of results, a row for each update}
  \item{melted.data}{melted results data frame for plotting}
  \item{correct}{number of variables detected correctly in each data set}
  \item{elapsed}{total elapsed time}
}
}
\description{
Create a data simulation and return train/holdout/test data sets.
}
\examples{
sim.type <- "sva"
num.samples <- 100
num.variables <- 100
pb <- 0.1
bias <- 0.4
nbias <- pb * num.variables
signals <- sprintf("gene\%04d", 1:nbias)
sim.data <- createSimulation(d=num.variables, n=num.samples, pb=pb,
                             bias=bias, type=sim.type, verbose=FALSE)
}
\seealso{
Other simulation: \code{\link{createDiffCoexpMatrixNoME}},
  \code{\link{simulateData}}
}
