% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{createSimulation}
\alias{createSimulation}
\title{Create a data simulation and return train/holdout/validation data sets.}
\usage{
createSimulation(num.samples = 100, num.variables = 100,
  pct.signals = 0.1, bias = 0.4, class.label = "class",
  sim.type = "mainEffect", pct.train = 0.5, pct.holdout = 0.5,
  pct.validation = 0, save.file = NULL, verbose = FALSE)
}
\arguments{
\item{num.samples}{An integer for the number of samples}

\item{num.variables}{An integer for the number of variables}

\item{pct.signals}{A numeric for proportion of simulated signal variables}

\item{bias}{A numeric for effect size in simulated signal variables}

\item{class.label}{A character vector for the name of the class column}

\item{sim.type}{A character vector of the type of simulation:
mainEffect/interactionErdos/interactionScalefree}

\item{pct.train}{A numeric percentage of samples to use for traning}

\item{pct.holdout}{A numeric percentage of samples to use for holdout}

\item{pct.validation}{A numeric percentage of samples to use for testing}

\item{save.file}{A filename or NULL indicating whether to save the simulations to file}

\item{verbose}{A flag indicating whether verbose output be sent to stdout}
}
\value{
A list with:
\describe{
  \item{train}{traing data set}
  \item{holdout}{holdout data set}
  \item{validation}{validation data set}
  \item{class.label}{the class label/column name}
  \item{signal.names}{the variable names with simulated signals}
  \item{elapsed}{total elapsed time}
}
}
\description{
Create a data simulation and return train/holdout/validation data sets.
}
\examples{
num.variables <- 100
num.samples <- 100
pct.signals <- 0.1
bias <- 0.4
sim.type <- "mainEffect"
sim.data <- createSimulation(num.samples=num.samples,
                             num.variables=num.variables,
                             pct.signals=pct.signals,
                             bias=bias,
                             sim.type=sim.type,
                             verbose=FALSE)
}
\seealso{
Other simulation: \code{\link{createInteractions}},
  \code{\link{createMainEffects}},
  \code{\link{splitDataset}}
}
