% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/privateEC.R
\docType{package}
\name{privateEC-package}
\alias{privateEC-package}
\title{Welcome to privateEC}
\description{
Package Algorithms
\describe{
  \item{Simulation}{simulate data sets: mainEffect/interactionErdos/interactionScalefree}
  \item{Classification}{privateEC and comparison algorithms}
  \item{Workflow}{functions for reproducing paper analysis pipelines}
}
}
\references{
Trang Le, W. K. Simmons, M. Misaki, B.C. White, J. Savitz, J. Bodurka, and
B. A. McKinney. “Differential privacy-based Evaporative Cooling feature selection
and classification with Relief-F and Random Forests,” Bioinformatics.
Accepted. https://doi.org/10.1093/bioinformatics/btx298. 2017.

This work has been supported in part by NIMH/NIH grants (R01 MH098099 to JB,
and K01 MH096175 to WKS).
}
