% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{privateEC}
\alias{privateEC}
\title{Private Evaporative Cooling feature selection and classification}
\usage{
privateEC(train.ds = NULL, holdout.ds = NULL, validation.ds = NULL,
  label = "phenos", is.simulated = TRUE, bias = 0.4, myrun = "001",
  update.freq = 50, corelearn.estimator = "ReliefFbestK",
  start.temp = 0.1, final.temp = 10^(-5), tau.param = 100,
  threshold = 4/sqrt(nrow(holdout.ds)),
  tolerance = 1/sqrt(nrow(holdout.ds)), signal.names = NULL,
  shortname = "paramstring", save.file = NULL, verbose = FALSE)
}
\arguments{
\item{train.ds}{A data frame with training data and class labels}

\item{holdout.ds}{A data frame with holdout data and class labels}

\item{validation.ds}{A data frame with validation data and class labels}

\item{label}{A character vector of the class variable column name}

\item{is.simulated}{Is the data simulated (or real?)}

\item{bias}{A numeric for effect size in simulated signal variables}

\item{myrun}{A character vector of a unique run identifier}

\item{update.freq}{An integer the number of steps before update}

\item{corelearn.estimator}{CORElearn Relief-F estimator}

\item{start.temp}{A numeric EC starting temperature}

\item{final.temp}{A numeric EC final temperature}

\item{tau.param}{A numeric tau to control temperature reduction schedule}

\item{threshold}{A numeric, default 4 / sqrt(n) suggested in the
thresholdout’s supplementary material (Dwork, et al.,2015)}

\item{tolerance}{A numeric, default 1 / sqrt(n) suggested in the
thresholdout’s supplementary material (Dwork, et al.,2015)}

\item{signal.names}{A character vector of signal names in simulated data}

\item{shortname}{A character vector of a parameters separated by '_'}

\item{save.file}{A character vector for results filename or NULL to skip}

\item{verbose}{A flag indicating whether verbose output be sent to stdout}
}
\value{
A list with:
\describe{
  \item{algo.acc}{data frame of results, a row for each update}
  \item{ggplot.data}{melted results data frame for plotting with ggplot}
  \item{correct}{number of variables detected correctly in each data set}
  \item{elapsed}{total elapsed time}
}
}
\description{
Private Evaporative Cooling feature selection and classification
}
\note{
Within thresholdout, we choose a threshold of 4 / sqrt(n) and
tolerance of 1 / sqrt(n) as suggested in the thresholdout’s supplementary
material (Dwork, et al., 2015).
}
\examples{
num.samples <- 100
num.variables <- 100
pct.signals <- 0.1
sim.data <- createSimulation(n=num.samples,
                             num.vars=num.variables,
                             pct.signals=pct.signals,
                             sim.type="mainEffect",
                             verbose=FALSE)
pec.results <- privateEC(train.ds=sim.data$train,
                         holdout.ds=sim.data$holdout,
                         validation.ds=sim.data$validation,
                         label=sim.data$class.label,
                         is.simulated=TRUE,
                         signal.names=sim.data$signal.names,
                         verbose=FALSE)
}
\references{
Trang Le, W. K. Simmons, M. Misaki, B.C. White, J. Savitz, J. Bodurka,
and B. A. McKinney. “Differential privacy-based Evaporative Cooling feature selection
and classification with Relief-F and Random Forests,”
Bioinformatics. Accepted. https://doi.org/10.1093/bioinformatics/btx298. 2017
}
\seealso{
Other classification: \code{\link{originalThresholdout}},
  \code{\link{privateECinbix}}, \code{\link{standardRF}}
}
