% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{privateECinbix}
\alias{privateECinbix}
\title{Call C++ inbix Evaporative Cooling Privacy}
\usage{
privateECinbix(train.ds = NULL, holdout.ds = NULL, validation.ds = NULL,
  label = "phenos", is.simulated = TRUE, bias = 0.4, myrun = "001",
  update.freq = 50, start.temp = 0.1, final.temp = 10^(-5),
  tau.param = 100, shortname = "paramstring", save.file = NULL,
  verbose = FALSE)
}
\arguments{
\item{train.ds}{A data frame with training data and class labels}

\item{holdout.ds}{A data frame with holdout data and class labels}

\item{validation.ds}{A data frame with validation data and class labels}

\item{label}{A character vector of the class variable column name}

\item{is.simulated}{Is the data simulated (or real?)}

\item{bias}{A numeric for effect size in simulated signal variables}

\item{myrun}{A character vector of a unique run identifier}

\item{update.freq}{A integer for the number of steps before update}

\item{start.temp}{A numeric for EC starting temperature}

\item{final.temp}{A numeric for EC final temperature}

\item{tau.param}{A numeric for tau to control reduction schedule}

\item{shortname}{A character vector of a parameters separated by '_'}

\item{save.file}{A character vector for results filename or NULL to skip}

\item{verbose}{A flag indicating whether verbose output be sent to stdout}
}
\value{
A list containing:
\describe{
  \item{algo.acc}{data frame of results, a row for each update}
  \item{ggplot.data}{melted results data frame for plotting}
  \item{correct}{number of variables detected correctly in each data set}
  \item{elapsed}{total elapsed time}
}
}
\description{
Assumes the inbix executable is in the PATH.
}
\note{
inbix must be instaled in the path
}
\examples{
num.samples <- 100
num.variables <- 100
pct.signals <- 0.1
sim.data <- createSimulation(num.vars=num.variables, n=num.samples,
                             sim.type="mainEffect", verbose=FALSE)
pec.results <- privateECinbix(train.ds=sim.data$train,
                              holdout.ds=sim.data$holdout,
                              validation.ds=sim.data$validation,
                              verbose=FALSE)
}
\seealso{
Other classification: \code{\link{originalThresholdout}},
  \code{\link{privateEC}}, \code{\link{standardRF}}
}
