% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{standardRF}
\alias{standardRF}
\title{Standard random forests algorithm serves as a baseline model}
\usage{
standardRF(data.sets = NULL, label = "phenos", is.simulated = TRUE,
  sim.type = "mainEffect", signal.names = NULL, shortname = "paramstring",
  save.file = NULL, verbose = FALSE)
}
\arguments{
\item{data.sets}{A list of train, holdout and validation data frames}

\item{label}{A character vector of the class variable column name}

\item{is.simulated}{Is the data simulated (or real?)}

\item{sim.type}{A character vector of the type of simulation:
mainEffect/interactionErdos/interactionScalefree}

\item{signal.names}{A character vector of signal names in simulated data}

\item{shortname}{A character vector of a parameters separated by '_'}

\item{save.file}{A character vector for results filename or NULL to skip}

\item{verbose}{A flag indicating whether verbose output be sent to stdout}
}
\value{
A list containing:
\describe{
  \item{plots.data}{data frame of results, a row for each update}
  \item{melted.data}{melted results data frame for plotting}
  \item{correct}{number of variables detected correctly in each data set}
  \item{elapsed}{total elapsed time}
}
}
\description{
Standard random forests algorithm serves as a baseline model
}
\examples{
num.samples <- 100
num.variables <- 100
pct.signals <- 0.1
nbias <- pct.signals * num.variables
signals <- sprintf("gene\%04d", 1:nbias)
data.sets <- createSimulation(d=num.variables, n=num.samples,
                              sim.type="mainEffect", verbose=FALSE)
rra.results <- standardRF(data.sets=data.sets,
                          is.simulated=TRUE,
                          type=type,
                          verbose=FALSE,
                          signal.names=signals)
}
\seealso{
Other classification: \code{\link{originalThresholout}},
  \code{\link{privateECinbix}}, \code{\link{privateEC}}
}
