% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{standardRF}
\alias{standardRF}
\title{Standard random forests algorithm serves as a baseline model}
\usage{
standardRF(train.ds = NULL, holdout.ds = NULL, validation.ds = NULL,
  label = "phenos", rf.ntree = 500, rf.mtry = NULL, is.simulated = TRUE,
  signal.names = NULL, save.file = NULL, verbose = FALSE)
}
\arguments{
\item{train.ds}{A data frame with training data and class labels}

\item{holdout.ds}{A data frame with holdout data and class labels}

\item{validation.ds}{A data frame with validation data and class labels}

\item{label}{A character vector of the class variable column name}

\item{rf.ntree}{An integer the number of trees in the random forest}

\item{rf.mtry}{An integer the number of variables sampled at each random forest node split}

\item{is.simulated}{Is the data simulated (or real?)}

\item{signal.names}{A character vector of signal names in simulated data}

\item{save.file}{A character vector for results filename or NULL to skip}

\item{verbose}{A flag indicating whether verbose output be sent to stdout}
}
\value{
A list containing:
\describe{
  \item{algo.acc}{data frame of results, a row for each update}
  \item{ggplot.data}{melted results data frame for plotting}
  \item{correct}{number of variables detected correctly in each data set}
  \item{elapsed}{total elapsed time}
}
}
\description{
Standard random forests algorithm serves as a baseline model
}
\examples{
num.samples <- 100
num.variables <- 100
pct.signals <- 0.1
sim.data <- createSimulation(num.variables = num.variables,
                             num.samples = num.samples,
                             sim.type = "mainEffect",
                             pct.train = 1 / 3,
                             pct.holdout = 1 / 3,
                             pct.validation = 1 / 3,
                             verbose = FALSE)
rra.results <- standardRF(train.ds=sim.data$train,
                          holdout.ds=sim.data$holdout,
                          validation.ds=sim.data$validation,
                          label=sim.data$class.label,
                          is.simulated=TRUE,
                          verbose=FALSE,
                          signal.names=sim.data$signal.names)
}
\seealso{
Other classification: \code{\link{originalThresholdout}},
  \code{\link{privateEC}}, \code{\link{privateRF}},
  \code{\link{xgboostRF}}
}
